susbsistema de registrar prenda final parte 1

+++ DTO_negocio/src/main/java/com/mycompany/dto_negocio/PagoDTO.java
@@ -10,12 +10,12 @@ package com.mycompany.dto_negocio;
  */
 public class PagoDTO {
     private String metodo;
-    private int monto;
+    private double monto;
     private boolean exitoso;
 
     public PagoDTO() {}
 
-    public PagoDTO(String metodo, int monto, boolean exitoso) {
+    public PagoDTO(String metodo, double monto, boolean exitoso) {
         this.metodo = metodo;
         this.monto = monto;
         this.exitoso = exitoso;
@@ -24,8 +24,8 @@ public class PagoDTO {
     public String getMetodo() { return metodo; }
     public void setMetodo(String metodo) { this.metodo = metodo; }
 
-    public int getMonto() { return monto; }
-    public void setMonto(int monto) { this.monto = monto; }
+    public double getMonto() { return monto; }
+    public void setMonto(double monto) { this.monto = monto; }
 
     public boolean isExitoso() { return exitoso; }
     public void setExitoso(boolean exitoso) { this.exitoso = exitoso; }
